+++ src/components/ui/GlobalSearch.tsx
@@ -1,52 +1,55 @@
-import { Loader, Menu, TextInput } from '@mantine/core';
-import { useField } from '@mantine/form';
-import { IconSearch } from '@tabler/icons-react';
-import { useGlobalSearch } from '../../api/common/queries';
-import { useDidUpdate } from '@mantine/hooks';
-import { useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
-import GlobalSearchItem from './GlobalSearchItem';
+import { Center, Combobox, Highlight, Stack, Text, Title } from '@mantine/core';
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import Search from './Search';
+import { Group } from '@mantine/core';
+import UserAvatar from './UserAvatar';
+import { useSearchUsers } from '../../api/user/queries';
+import { IUser } from '../../types/models';
 
 export default function GlobalSearch() {
-  const query = useField({
-    initialValue: '',
-  });
-  const { data, isLoading, isFetching, refetch } = useGlobalSearch(query.getValue());
-  const location = useLocation();
-
-  useDidUpdate(() => {
-    refetch();
-  }, [query.getValue()]);
-
-  useEffect(() => {
-    query.reset();
-  }, [location.pathname]);
+  const [query, setQuery] = useState('');
+  const { data, isLoading, isFetching, refetch } = useSearchUsers(query);
+  const navigate = useNavigate();
 
   return (
-    <Menu opened={data?.length !== 0}>
-      <Menu.Target>
-        <TextInput
-          styles={(theme) => ({
-            input: {
-              boxShadow: theme.shadows.xs,
-            },
-          })}
-          w={200}
-          size='xs'
-          placeholder='Search...'
-          leftSection={<IconSearch size={15} />}
-          rightSection={isLoading || isFetching ? <Loader size={12} /> : null}
-          {...query.getInputProps()}
-        />
-      </Menu.Target>
-
-      {data && data.length > 0 && (
-        <Menu.Dropdown w='min(90%, 30rem)' mt={10}>
-          {data.map((item) => (
-            <GlobalSearchItem key={item.id} item={item} query={query.getValue()} />
-          ))}
-        </Menu.Dropdown>
+    <Search
+      isFetching={isFetching}
+      isLoading={isLoading}
+      query={query}
+      setQuery={setQuery}
+      refetch={refetch}
+      data={data}
+      comboboxProps={{ withinPortal: true }}
+      inputProps={{
+        size: 'xs',
+      }}
+      onSubmit={function (value, store) {
+        navigate(value);
+        setQuery('');
+        store.closeDropdown();
+      }}
+    >
+      {({ item }) => (
+        <Combobox.Option value={`/profile/${item.username}`} key={item.id}>
+          <Group>
+            <Center w={40}>
+              <UserAvatar size='md' user={item as IUser} color='yellow' variant='filled' />
+            </Center>
+            <Stack gap={0}>
+              <Title order={5}>
+                <Highlight
+                  fz='sm'
+                  highlight={query}
+                >{`${item.first_name} ${item.last_name}`}</Highlight>
+              </Title>
+              <Text component='span' c='dimmed' fz='xs'>
+                <Highlight fz='xs' highlight={query}>{`@${item.username}`}</Highlight>
+              </Text>
+            </Stack>
+          </Group>
+        </Combobox.Option>
       )}
-    </Menu>
+    </Search>
   );
 }